made some optional functionality so only users can edit or delete articles they added, as well as can only see articles they've added

+++ src/components/articles/Article.js
@@ -1,12 +1,14 @@
 import { Link } from "react-router-dom"
 
-export const Article = ({ id, title, synopsis, url, getAllArticles }) => {
-    return <section className="article">
-        <div>
-            <a href={url} target="_blank"><h5>{title}</h5></a>
-        </div>
-        <div className="synopsis">Synopsis: <em>{synopsis}</em></div>
-        <button
+export const Article = ({ id, title, synopsis, url, getAllArticles, articleObject }) => {
+
+   const localHoneyUser = localStorage.getItem("honey_user")
+    const honeyUserObject = JSON.parse(localHoneyUser)
+
+    const buttonOrNoButton = () => {
+        if (honeyUserObject.id === articleObject.userId) {
+            return<>
+             <button
             onClick={() => {
                 fetch(`http://localhost:8088/articles/${id}`, {
                     method: "DELETE"
@@ -17,5 +19,19 @@ export const Article = ({ id, title, synopsis, url, getAllArticles }) => {
             }}
         >Delete</button>
         <Link to={`/articles/${id}`}><button>Edit</button></Link>
+        </>
+        } 
+        else {
+            return ""
+        }
+    }
+
+
+    return <section className="article">
+        <div>
+            <a href={url} target="_blank"><h5>{title}</h5></a>
+        </div>
+        <div className="synopsis">Synopsis: <em>{synopsis}</em></div>
+        {buttonOrNoButton()}
     </section>
 }
